Add files via upload
This file is used for saving the image files that the computer reads from the ROS Topic image_data. It calls the files that convert the images to CV2 files, read the meta data of the images including the location of the robot when the picture was taken, and saves the files in the correct python directory.